+++ R/users.R
@@ -8,21 +8,15 @@
 #' @examples
 #'
 #' \dontrun{
-#'
-#' ## select one or more twitter users to lookup
 #' users <- c(
 #'   "potus", "hillaryclinton", "realdonaldtrump",
 #'   "fivethirtyeight", "cnn", "espn", "twitter"
 #' )
-#' usr_df <- lookup_users(users)
+#' users <- lookup_users(users)
+#' users
 #'
-#' ## view tweet data for these users via tweets_data()
-#' tweets_data(usr_df)
-#' 
-#' # Find user data for the recent posters to #rstats
-#' rs <- search_tweets("#rstats", n = 500)
-#' user_ids <- unique(rs$user_id)
-#' lookup_users(user_ids)
+#' # latest tweet from each user
+#' tweets_data(users)
 #' }
 #'
 #' @return A tibble of users data.
@@ -48,5 +42,6 @@ lookup_users <- function(users, parse = TRUE, token = NULL,
   if (parse) {
     results <- users_with_tweets(results)
   }
+  
   results
 }